add routes and font-end files
