+++ application/other_files/db_setup.py
@@ -6,7 +6,7 @@ import os
 
 BASE_DIR=os.path.dirname(os.path.realpath(__file__))
 
-connection_string="sqlite:///"+os.path.join(BASE_DIR, 'site.db')
+connection_string="sqlite:///"+os.path.join(BASE_DIR, '../../site.db')
 
 Base=declarative_base()
