+++ application/static/styles.css
@@ -1,0 +1,120 @@
+ :root {
+  --light-blue: rgb(81, 208, 254);
+  --gray: rgb(176, 179, 179);
+}
+
+body {
+  width: 90%;
+  margin: auto;
+  font-family: "Raleway", sans-serif;
+  background-color: #ffffff;
+  box-sizing: border-box;
+  font-size: 16px;
+}
+
+/* --------- General styling ------------ */
+.small-logo {
+  height: 3rem;
+  width: 3rem;
+}
+
+.logo {
+  height: 5rem;
+  width: 5rem;
+}
+
+
+
+/* --------- Home.html styling ------------ */
+/* Container needed to position the button. Width auto so it aligns with nav */
+.container-home {
+  position: absolute;
+  width: auto;
+  padding-top: 50px;
+  padding-bottom: 100px;
+}
+
+.container-home #logo-banner {
+  position: relative;
+  margin: auto;
+  width: 90%;
+  padding-bottom: 50px;
+}
+
+.container-home .h1 {
+    font-weight: bold;
+    font-size: 2vw;
+}
+
+.container-home .new-customer {
+   padding-top: 50px;
+  padding-bottom: 30px;
+  width: 100%;
+  margin: auto;
+
+}
+
+.container-home .existing-customer {
+  padding-bottom: 50px;
+  width: 100%;
+  margin: auto;
+
+}
+
+/* homepage button styling */
+.container-home .btn1 {
+  position: relative;
+  margin-top: 30px;
+   top: 80%;
+  left: 30%;
+    transform: translate(-20%, -20%);
+  -ms-transform: translate(-50%, -50%);
+    background-color: #3AB4FC;
+  color: white;
+  font-size: 20px;
+  padding: 12px 24px;
+  border: none;
+  cursor: pointer;
+  border-radius: 5px;
+  }
+
+.container-home .btn2 {
+    margin-top: 30px;
+  position: relative;
+   top: 80%;
+  left: 30%;
+    transform: translate(-20%, -20%);
+  -ms-transform: translate(-50%, -50%);
+    background-color: #3AB4FC;
+  color: white;
+  font-size: 20px;
+  padding: 12px 24px;
+  border: none;
+  cursor: pointer;
+  border-radius: 5px;
+  }
+
+/* Javascript to make home page buttons change colour when interacted */
+.container-home .btn1:hover {
+  background-color: #D3D3D3;
+}
+
+.container-home .btn2:hover {
+  background-color: #D3D3D3;
+}
+
+/* --------- Currency.html styling ------------ */
+.currency-elements .gray-box {
+  width: 85%;
+}
+
+/* --------- Footer styling ------------ */
+
+footer {
+  font-size: 1 em;
+  text-align: center;
+  color: white;
+  font-family: "Raleway", sans-serif;
+  background-color: #036aa9;
+}
+
