+++ application/templates/currency.html
@@ -1,0 +1,25 @@
+{% extends "layout.html" %}
+{% block body_content %}
+
+<!--<&#45;&#45;! start html code from here - no need for the boilerplate code as this is inherited from layout.html &ndash;&gt;-->
+
+
+<section class="d-flex flex-column justify-content-center align-items-center h-100 currency-elements">
+    <form class="gray-box d-flex flex-column btn active">
+        <label for="GBPamount">How much money do you want to withdraw?</label>
+        <input type="GBPamount" id="fname" name="GBPamount">
+        
+        <label for="currency_select">Choose currency:</label>
+        <select class="form-select form-select-sm" id="currency_select">
+            {% for i in all_currencies %}
+                <option value="{{i[0]}}">{{i[0]}} - {{i[1]}} - {{i[2]}}</option>
+            {%endfor%}
+        </select>
+        <a class="blue-box  btn btn-lg active" role="button" href="#">
+            Buy currency! 
+        </button>
+    </form>
+</section>
+
+<!--<&#45;&#45;! all html code should be before the below endblock tag so the footer remains at the bottom of page &ndash;&gt;-->
+{% endblock %}
