+++ application/templates/register.html
@@ -1,0 +1,73 @@
+{% extends "layout.html" %}
+{% block body_content %}
+<div class="forms">
+    <form action="", method="POST">
+        {{ form.hidden_tag() }}
+        {{ form.csrf_token() }}
+        <fieldset>
+            <legend>Sign up Today</legend>
+
+
+            <div class="form-group">
+                {{ form.first_name.label(class="form-control-label") }}
+                {{ form.first_name(class="form-control form-control-lg", placeholder="First name") }}
+            </div>
+            <div class="form-group">
+                {{ form.last_name.label(class="form-control-label") }}
+                {{ form.last_name(class="form-control form-control-lg", placeholder="Last name") }}
+            </div>
+            <div class="form-group">
+                {{ form.email.label(class="form-control-label") }}
+                {{ form.email(class="form-control form-control-lg", placeholder="example@example.com") }}
+                <ul>
+                    {% for error in form.email.errors %}
+                        <li style="color:red;">{{ error }}</li>
+                    {% endfor %}
+                </ul>
+            </div>
+            <div class="form-group">
+                {{ form.username.label(class="form-control-label") }}
+                {{ form.username(class="form-control form-control-lg", placeholder="Your username here") }}
+                 <ul>
+                    {% for error in form.username.errors %}
+                        <li style="color:red;">{{ error }}</li>
+                    {% endfor %}
+                </ul>
+            </div>
+            <div class="form-group">
+                {{ form.password.label(class="form-control-label") }}
+                {{ form.password(class="form-control form-control-lg", placeholder="Password") }}
+            </div>
+            <div class="form-group">
+                {{ form.confirm_password.label(class="form-control-label") }}
+                {{ form.confirm_password(class="form-control form-control-lg", placeholder="Repeat your password") }}
+                <ul>
+                {% for error in form.confirm_password.errors %}
+                        <li style="color:red;">{{ error }}</li>
+                    {% endfor %}
+                </ul>
+            </div>
+            <div class="form-group">
+                {{ form.address.label(class="form-control-label") }}
+                {{ form.address(class="form-control form-control-lg", placeholder="Address line 1") }}
+            </div>
+            <div class="form-group">
+                {{ form.postcode.label(class="form-control-label") }}
+                {{ form.postcode(class="form-control form-control-lg", placeholder="Postcode") }}
+            </div>
+
+        </fieldset>
+        <div class="form-group">
+            {{ form.submit(class="btn btn-secondary") }}
+        </div>
+
+    </form>
+    <div class="border-top pt-3">
+        <small class="text-muted">
+            Already Have An Account? <a class="ml-2" href="{{ '/login' }}">Sign In</a>
+        </small>
+
+</div>
+</div>
+
+{% endblock %}
