+++ application/templates/transactions.html
@@ -1,0 +1,16 @@
+{% extends "layout.html" %}
+{% block body_content %}
+
+<--! start html code from here - no need for the boilerplate code as this is inherited from layout.html -->
+
+
+
+
+
+
+
+
+
+
+<--! all html code should be before the below endblock tag so the footer remains at the bottom of page -->
+{% endblock %}
