+++ requirements.txt
@@ -1,2 +1,5 @@
 flask
 flask_sqlalchemy
+flask_wtf
+wtforms
+wtforms.validators
